Exercise - 2
Write HTML and CSS code to style a paragraph inside a div which contains few other para. The first paragraph must have background-color as yellow and text-color red. The other paragraph must have background-color as blue and text-color as white . We don't have to use id and classes for paragraph's and div's .